remove duplicated code & use different import for secp256k1
