+++ testApp/go.mod
@@ -4,9 +4,8 @@ go 1.13
 
 require (
 	github.com/ElrondNetwork/elrond-go v1.1.2
-	github.com/btcsuite/btcutil v1.0.2
-	github.com/decred/dcrd/dcrec/secp256k1 v1.0.3
-	github.com/ethereum/go-ethereum v1.10.4
+	github.com/btcsuite/btcd v0.22.0-beta
+	github.com/btcsuite/btcutil v1.0.3-0.20201208143702-a53e38424cce
 	github.com/karalabe/hid v1.0.0
 	golang.org/x/crypto v0.0.0-20210322153248-0c34fe9e7dc2
 )
